api-frontend/backend/requirements.txt
kafka-python
pyspark
flask
hdfs==2.6.0
flask_cors
